+++ src/finance/invoice/operations/Operate.java
@@ -1,8 +1,8 @@
 package finance.invoice.operations;
 
 public class Operate {
-	public Integer multipile(int data) {
-		int out = data * 5000;
+	public float multipile(float data) {
+		float out = data * 5000;
 		return out;
 	}
 }
